[graal] Change logic to create worktree
This code changes the logic to create the worktree.
The default branch is used to initialize the worktree,
if no branch is passed as input. Otherwise the latter
is used to initialize the worktree.

This change is needed to handle repositories not
having the branch `master` or which default branch
is not `master`.

Tests have been added accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>