[travis] Change how to install nomos
This code changes the way how to install the nomos. Thus,
instead of creating the executable, it relies on the deb
packages fossology-common and fossology-nomos.

Due to the missing dependency libjson-c3 in the defaul distribution
used by Travis (`xenial`), the distribution has been set to `bionic`.

The nomossa executable path has been updated accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>