[travis] Update .travis.yml
This code aims at aligning the CI tests across the
different grimoirelab components.

Python 3.5 is removed as its support period has ended and
Python 3.6, 3.7 and 3.8 are now supported.

The version of setuptools and pip has been downgraded as a
hotfix to solve the failing CI tests.

The version of Go is updated to the most stable version to
align it with the scc tool requirement.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>